pyproject.toml
[tool.poetry]
name = "picnik"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.103.2"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
pylint = "^3.0.1"
pytest = "^7.4.3"
prisma = "^0.11.0"
pydantic = {extras = ["email"], version = "^2.4.2"}
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
python-multipart = "^0.0.6"
pillow = "^10.1.0"
arrow = "^1.3.0"
insightface = "^0.7.3"
onnxruntime = "^1.16.3"
typesense = "^0.18.0"
geopy = "^2.4.1"

[tool.poetry.scripts]
dev = 'uvicorn api.main:app --reload'
prisma-push = 'prisma db push'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "."
]
